feat(contexts): adds support for passing a context of either 'window', 'worker', or 'node' as a query parameter
Some polyfills only work in specific environments such as where the 'window' object is available.
Other polyfills work in several environments, but ship different versions for different environments.
This commit adds a new (optional) query parameter, 'context', which can be either 'window', 'worker', or 'node'.
If it isn't given, it defaults to 'window'. When a set of polyfills is requested, those that doesn't support the given
context won't be returned. For example, if you request 'request-animation-frame', and pass in 'worker' as the context,
the polyfill won't be added, even if you also pass in the 'force' option. Fixes #2